+++ test/smoke/SmokeConfig.ts
